refactor: update filenames for views
Update view filenames to reflect new class names for views in app
